net/tcp: Move net_tcp_set_hdr() away from net core
Only the unit test needs it now, so let's move it there.
