+++ subsys/net/ip/net_core.c
@@ -26,7 +26,6 @@ LOG_MODULE_REGISTER(net_core, CONFIG_NET_CORE_LOG_LEVEL);
 #include <net/net_pkt.h>
 #include <net/net_core.h>
 #include <net/dns_resolve.h>
-#include <net/tcp.h>
 #include <net/gptp.h>
 #include <net/lldp.h>
