+++ subsys/net/ip/tcp.c
@@ -24,7 +24,6 @@ LOG_MODULE_REGISTER(net_tcp, CONFIG_NET_TCP_LOG_LEVEL);
 #include <net/net_pkt.h>
 #include <net/net_ip.h>
 #include <net/net_context.h>
-#include <net/tcp.h>
 #include <misc/byteorder.h>
 
 #include "connection.h"
@@ -1264,41 +1263,6 @@ bool net_tcp_validate_seq(struct net_tcp *tcp, struct net_tcp_hdr *tcp_hdr)
 		(net_tcp_seq_cmp(sys_get_be32(tcp_hdr->seq),
 				 tcp->send_ack
 					+ net_tcp_get_recv_wnd(tcp)) < 0);
-}
-
-struct net_tcp_hdr *net_tcp_set_hdr(struct net_pkt *pkt,
-				    struct net_tcp_hdr *hdr)
-{
-	NET_PKT_DATA_ACCESS_DEFINE(tcp_access, struct net_tcp_hdr);
-	struct net_pkt_cursor backup;
-	struct net_tcp_hdr *tcp_hdr;
-	bool overwrite;
-
-	overwrite = net_pkt_is_being_overwritten(pkt);
-	net_pkt_set_overwrite(pkt, true);
-
-	net_pkt_cursor_backup(pkt, &backup);
-	net_pkt_cursor_init(pkt);
-
-	if (net_pkt_skip(pkt, net_pkt_ip_hdr_len(pkt) +
-			 net_pkt_ipv6_ext_len(pkt))) {
-		tcp_hdr = NULL;
-		goto out;
-	}
-
-	tcp_hdr = (struct net_tcp_hdr *)net_pkt_get_data_new(pkt, &tcp_access);
-	if (!tcp_hdr) {
-		goto out;
-	}
-
-	memcpy(tcp_hdr, hdr, sizeof(struct net_tcp_hdr));
-
-	net_pkt_set_data(pkt, &tcp_access);
-out:
-	net_pkt_cursor_restore(pkt, &backup);
-	net_pkt_set_overwrite(pkt, overwrite);
-
-	return tcp_hdr == NULL ? NULL : hdr;
 }
 
 int net_tcp_finalize(struct net_pkt *pkt)
